fix(dashboard): server-side auth guards + error boundaries for admin & staff
Phase 1 of the dashboard-architecture unification: bring the two laggard
trees (admin, staff) up to the org/org-workspace flagship baseline on the
two axes where they were weakest — access control and error handling.

Access control (was client-side only):
- admin/layout.tsx and staff/[staffId]/layout.tsx are now thin SERVER
  components. Admin runs requireUserRole("ADMIN"); staff runs
  requireOnboarded() + the existing ADMIN|STAFF|matching-staffProfileId
  rule — both enforced before any markup ships to the client, replacing
  the previous useSession()+useQuery(/api/user)+post-hydration router.replace
  that leaked the protected shell to unauthorized users before bouncing them.
- Chrome moves to client shells (AdminShell / StaffShell) that receive the
  server-resolved identity as props (no useSession first-render null → no
  hydration mismatch on the sidebar name, no /api/user fetch waterfall).
- Sign-out consolidated onto the shared signOutEverywhere() util.

Error handling (was zero route boundaries):
- Add admin/error.tsx and staff/[staffId]/error.tsx as thin wrappers over
  DashboardRouteError (Sentry capture + digest + reset), matching org /
  consultant / consultee. Staff page content also gains DashboardErrorBoundary.

Visual output preserved (same sidebar + sticky top bar). tsc clean.
